pyproject.toml
[project]
name = "cline-core"
version = "0.0.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Adem Gaygusuz", email = "adem@ardweb.co.uk" }
]
requires-python = ">=3.12"

[build-system]
requires = ["uv>=0.4.0", "grpcio-tools>=1.64.0"]

[dependency-groups]
dev = [
    "build",
    "grpcio-tools>=1.64.0",
    "grpcio>=1.64.0",
    "pytest>=9.0.1",
    "python-semantic-release>=10.5.2",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
allow_zero_version = true
major_on_zero = false
